+++ packages/amplify-category-function/src/provider-utils/awscloudformation/utils/functionPluginLoader.ts
@@ -26,7 +26,8 @@ export async function templateWalkthrough(context: any, params: Partial<Function
       return (
         condition.provider === params.providerContext.provider &&
         condition.service === params.providerContext.service &&
-        (condition.runtime === params.runtime.value || condition.runtime.includes(params.runtime.value))
+        (condition.runtime === params.runtime.value ||
+          (Array.isArray(condition.runtime) && condition.runtime.includes(params.runtime.value)))
       );
     },
     selectionPrompt: 'Choose the function template that you want to use:',
@@ -140,7 +141,7 @@ async function getSelectionFromContributors<T>(context: any, selectionOptions: P
   };
 }
 
-async function loadPluginFromFactory(pluginPath, expectedFactoryFunction, context): Promise<any> {
+export async function loadPluginFromFactory(pluginPath, expectedFactoryFunction, context): Promise<any> {
   let plugin;
   try {
     plugin = await import(pluginPath);