+++ packages/amplify-nodejs-runtime-provider/src/utils/invokeOptions.ts
@@ -1,0 +1,7 @@
+export type InvokeOptions = {
+  packageFolder: string;
+  handler: string;
+  event: string;
+  context?: object;
+  environment?: { [key: string]: string };
+};