+++ packages/amplify-provider-awscloudformation/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.17.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-provider-awscloudformation@4.17.0...amplify-provider-awscloudformation@4.17.1) (2020-03-26)
+
+**Note:** Version bump only for package amplify-provider-awscloudformation
+
+
+
+
+
 # [4.17.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-provider-awscloudformation@4.16.1...amplify-provider-awscloudformation@4.17.0) (2020-03-22)